+++ src/renderer/components/PlayingView/TimeProgressBar.vue
@@ -30,7 +30,7 @@
         :widthOfThumbnail="widthOfThumbnail"
         :heightOfThumbnail="heightOfThumbnail"
         :screenshotContent="screenshotContent"
-        :currentTime="thumbnailCurrentTime"/> -->
+        :currentTime="thumbnailCurrentTime"/>
         <!-- translate优化 -->
       <div class="progress-ready" ref="readySlider">
         <div class="background-line"></div>
@@ -106,7 +106,7 @@ export default {
   methods: {
     appearProgressSlider() {
       this.isOnProgress = true;
-      this.$bus.$emit('clearAllWidgetDisappearDelay');
+      this.$bus.$emit('clear-all-widget-disappear-delay');
       this.$refs.playedSlider.style.height = PROGRESS_BAR_HEIGHT;
       this.$refs.readySlider.style.height = PROGRESS_BAR_HEIGHT;
       this.$refs.foolProofBar.style.height = PROGRESS_BAR_HEIGHT;
@@ -250,7 +250,7 @@ export default {
     },
     $_hideAllWidgets() {
       this.timeoutIdOfHideAllWidgets = setTimeout(() => {
-        this.$bus.$emit('hideAllWidgets');
+        this.$bus.$emit('hide-all-widgets');
       }, 3000);
     },
   },
